+++ lib/index.d.ts
@@ -1,6 +1,6 @@
-export * as near from "./api";
-export * from "./types";
-export * from "./near-bindgen";
 export * from "./collections";
-export * from "./utils";
+export * from "./types";
+export * as near from "./api";
+export * from "./near-bindgen";
 export * from "./promise";
+export * from "./utils";
