+++ lib/types/primitives.js
@@ -1,1 +1,8 @@
-export {};
+/**
+ * One yoctoNEAR. 10^-24 NEAR.
+ */
+export const ONE_YOCTO = 1n;
+/**
+ * One NEAR. 1 NEAR = 10^24 yoctoNEAR.
+ */
+export const ONE_NEAR = 1000000000000000000000000n;
